esplora: move builder to module and implement it for blocking too
